[tests] Move matrix tests to TestMatrices
