+++ assembly/src/bin/tornado-test.py
@@ -43,7 +43,6 @@ class TestEntry:
 __TEST_THE_WORLD__ = [
     TestEntry("uk.ac.manchester.tornado.unittests.TestHello"),
     TestEntry("uk.ac.manchester.tornado.unittests.arrays.TestArrays"),
-    TestEntry("uk.ac.manchester.tornado.unittests.arrays.TestMatrix"),
     TestEntry("uk.ac.manchester.tornado.unittests.functional.TestLambdas"),
     TestEntry("uk.ac.manchester.tornado.unittests.vectortypes.TestFloats"),
     TestEntry("uk.ac.manchester.tornado.unittests.vectortypes.TestDoubles"),
